feat: Moved model for get sticker pack.
