+++ botx/client/stickers_api/get_sticker_pack.py
@@ -1,13 +1,10 @@
-from typing import List, Optional
 from uuid import UUID
-
-from typing_extensions import Literal  # For python 3.7 support
 
 from botx.client.authorized_botx_method import AuthorizedBotXMethod
 from botx.client.botx_method import response_exception_thrower
 from botx.client.stickers_api.exceptions import StickerPackOrStickerNotFoundError
-from botx.models.api_base import UnverifiedPayloadBaseModel, VerifiedPayloadBaseModel
-from botx.models.stickers import Sticker, StickerPack
+from botx.client.stickers_api.sticker_pack import BotXAPIGetStickerPackResponsePayload
+from botx.models.api_base import UnverifiedPayloadBaseModel
 
 
 class BotXAPIGetStickerPackRequestPayload(UnverifiedPayloadBaseModel):
@@ -19,44 +16,6 @@ class BotXAPIGetStickerPackRequestPayload(UnverifiedPayloadBaseModel):
         sticker_pack_id: UUID,
     ) -> "BotXAPIGetStickerPackRequestPayload":
         return cls(sticker_pack_id=sticker_pack_id)
-
-
-class BotXAPIGetStickerResult(VerifiedPayloadBaseModel):
-    id: UUID
-    emoji: str
-    link: str
-
-
-class BotXAPIGetStickerPackResult(VerifiedPayloadBaseModel):
-    id: UUID
-    name: str
-    public: bool
-    # TODO Fix after backend
-    stickers_order: Optional[List[UUID]]
-    stickers: List[BotXAPIGetStickerResult]
-
-
-class BotXAPIGetStickerPackResponsePayload(VerifiedPayloadBaseModel):
-    status: Literal["ok"]
-    result: BotXAPIGetStickerPackResult
-
-    def to_domain(self) -> StickerPack:
-        if self.result.stickers_order:
-            self.result.stickers.sort(
-                key=lambda pack: self.result.stickers_order.index(  # type:ignore
-                    pack.id,
-                ),
-            )
-
-        return StickerPack(
-            id=self.result.id,
-            name=self.result.name,
-            is_public=self.result.public,
-            stickers=[
-                Sticker(id=sticker.id, emoji=sticker.emoji, image_link=sticker.link)
-                for sticker in self.result.stickers
-            ],
-        )
 
 
 class GetStickerPackMethod(AuthorizedBotXMethod):
